--- mailService/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "mail-service",
-  "version": "1.0.0",
-  "type": "module",
-  "description": "",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.21.1",
-    "express": "^4.17.1",
-    "mysql2": "^2.2.5",
-    "sequelize": "^6.6.2",
-    "ts-node": "^10.0.0",
-    "typescript": "^4.3.2",
-    "uuid": "^8.3.2"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.12",
-    "@types/node": "^15.12.4",
-    "@types/uuid": "^8.3.0"
-  }
-}
